pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.0",
    "black>=25.1.0",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "notebook>=7.4.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "torch>=2.2",
    "transformers>=4.55.2",
]
